ProcessWrapper: Add received signals for stderr and stdout
These essentially are what the current signals used to do, but as they
now contain the full output, rather than just what was most recently
received, we also need these to happen. So, new pair of signals, so we
can (in particular) sniff stderr output from places where that's
important.