ClipAudioSource: Add equaliser/compressor, and sketchpadTrack property
The former are the logic which holds the actual equaliser and compressor
functionality, alongside a finaliseProcess call which gets used to apply
those two (when enabled), and then copy the result onto the output
buffers. This ensures that we only have to do the work once per sound,
though at the cost of individual per-voice control over the settings.
However, we will have to live with that, and we also have the internal
filter in SamplerSynthVoice to use for per-voice filtering. Less
powerful, but more performant, so that'll do.

The sketchpadTrack property is simply used to store information on which
sketchpad track an individual CAS is associated with. By default this
will be the global track, but it can be changed at runtime to match
another track (though we don't currently do that, except for the
technicality that while loading, we first register the CAS, and then
change its sketchpad track).