ClipAudioSource: Precalc pitch shift ratio (and simplify clip access)
The former ensures that we can use the pitch shift amount for live pitch
shifting without having to do any unnecessary calculations (which in
turn should hopefully allow us to pitch things around in interesting
ways when we get far enough as to have automation).

The latter is more of a simplification - there's no need to do all that,
when we know we've got the clip in hand.