SamplerSynthVoice: Feed SoundTouch before we start trying to use it
This ensures that the latency is kept in check. What this in reality
means is that, when using time stretching, pitch shifts will be delayed
by the soundTouch processing latency window (which is a few ms), which
is imperfect, but we'll have to live with. For perfect pitch shift
timing, use unstretched pitching instead.