ClipAudioSource: Add sample-unit logic for start,length,loopDelta
This allows us to be considerably more precise when doing the work.
SamplerSynthVoice will be using this for its operation, but also this
means we can set the positions with an extremely high precision,
allowing us to do fun things like essentially treat a ClipAudioSource
set to a looping playback mode essentially as a wavetable. Obviously not
implemented yet (that'll need some ui and other assistance), but that'll
be next.