ClipAudioSource: Add Wavetable PlaybackStyle, move a bunch of logic to c++
The first is primarily a visual style, which instructs the UI to
interact with the CAS in a certain way, to support a wavetable style
playback solution (that is, split into super-narrow bands, and move
through them, though we'll also allow for arbitrary window sizes and
positions, since really it's just a looping playback style).

This also adds a number of new properties, which contain logic
previously housed in python. This allows us to treat ClipAudioSource as
the single source of truth for the values that relate to it.