QPainterContext: Add basic support for fillPath
The current implementation doesn't cubic or quadratic to anything, it
just draws direct lines. However, since what we're using it for is,
essentially, just the WaveFormItem, and the waveforms are straight-line
paths, well... headache for another day.

Also some rejigging of the code to reduce early return stuff and just...
make it read more readishly ;)