WaveFormItem: Switch thumbnails to high-quality tracktion ones
This makes our thumbnails way, way nicer to look at when zoomed in (and
also kind of in general). There's more to be done for that for recording
previews, which also will need switching over, but this is a good start.

The snippet left in for clip:/ stuff is intended to eventually let us
just pass a clip ID to the item, instead of the filename. However, not
sure it'll really gain us a huge amount, as apparently the thumbnail
logic will always read the file anyway. Seems wasteful, not sure if
perhaps a more invasive bit of surgery would be needed to fix that.