CASSubvoiceSettings: New mini-container for subvoice settings
This is a simple data container which just manages the settings for an
individual subvoice as pertaining to a ClipAudioSource. The sub-voices
are technically full voices, but they are launched alongside the prime
voice (and stopped accordingly), which have adjustments made for pan,
pitch, and gain. This allows us to create more proper chorus and unison
style things, without having to occupy multiple sample slots on a track
with the same sample and make an all-matching keyzone setup, leaving
that to allow for the actual voicing logic that's supposed to allow for.