SamplerSynthVoice: Add support for the subvoice settings (also minor refactor)
We now apply (a precalculated version of) each subvoice's pitch, pan,
and gain amounts, if there is a subvoice to do so with.

The refactor is really just moving some logic around in the startNote
function, ensuring that we're not doing work we don't have to. It's kind
of not that important, i guess, but order of activities is a thing, and
there's no reason to do things we don't have to.