SamplerSynthVoice: Don't apply aftertouch etc during tailoff
This will happen because MPE controllers will send out a channel
aftertouch 0 after notes are ended as a part of the "reset channel and
clear for next note" logic. As such, we should not be using that
information here, rather that is essentially intended for the next
voice on that channel.