SamplerSynthVoice: Throttle gain changes a bit by ramping them
This probably really wants to be on more of a curve than a linear
change, but it already helps a huge amount (much less crackling when
moving fingers around on an mpe controller). And now we've got somewhere
for the change logic to go, so it'll be simpler to tweak later.