SamplerSynthVoice: Switch to live time-stretching for tempo as well
As we are already doing live timestretching for our pitch shifting, why
not also do the same for the tempo adjustments. This also means that we
can do it very-extremely-live, as in during playback, and still stay in
time. There is some amount of glitchiness when doing so (as we're
clearing the buffer and doing a bunch of recalculating when doing so),
but it works, which it didn't before.