SamplerSynthVoice: Don't over-fill the SoundTouch buffer
This also (at least temporarily) removes the buffer clearing logic,
but... might need to reintroduce that, to avoid ending up out of sync.
Need to test that further. However, for now, it only keeps directly
ahead of the required amount of samples, rather than waiting for the
buffer to say it's empty.