SamplerSynthVoice: Rip out live timestretching (too heavy)
Basically, we can do one stream without much trouble, but multiple
streams is way, way too much work for the cpu on even an rpi5. So, rip
that back out again, preparing to switch to a more tightly controlled
offline timestretching done by SamplerSynthSound.

Also sneaked in a little fix for the playback position logic, as well as
a playback rate thing (which is important for samples that are other
than 48KHz, which, well, is most samples coming from places other than
us).