SamplerSynthSound: Add threaded timestretching processing
This replaces the attempt at live timestretching (which, while it did
work, was much too heavy to run on more than one stream at a time). If
SoundTouch operated per-sample instead of in batches, we'd be able to do
this, but it does not, and i gather it reasonably would end up with
SoundTouch being much heavier in general if it did so as well, so, we'll
just have to deal with that.

However, offline timestretching works fine anyway, and as we now control
it in a much more strict fashion, we can still change the bpm live, and
end up with something which is reasonably close to solid playback. It
is not instant the way it was previously, as the timestretcher has to
actually process the sample, but it does still work. Automation will not
work as tightly as we might like, but, it'll still end up where at the
correct stretch rate and playback position and such.