PatternModel: Add explicit quantizing logic to live midi recording
This replaces the pseudo-magical code we've used since it was first
implemented. The new code is much more precise, and will lock the note
entries' start and end points to a grid of a certain size, regardless of
the length of the steps themselves (though quantize-to-step is still the
default, just that it is now actually properly quantized). The
quantizing is done by rounding to the nearest grid point.