SequenceModel: Switch to autosave location, add import/export functions
The first of these two will result in any existing sketchpads loading as
empty. To remedy this, and fix existing sketchpads, "just" move the
global folder into a subfolder of the sequences dir named "autosave" and
things should start working fine again.

The second part is to allow for simpler importing and exporting of
sequence models (such as is required by the sketchpad versioning logic)