PatternModel: Introduce a performance clone, and skeleton for nudge
The performance clone concept allows us to hold a temporary model used
internally for playback and storage, and which is, essentially, a
temporary clone of the entire contents of the pattern, which can be
either discarded once a performance is complete, or applied back onto
the model if so desired.

Initially we add some basic nudge functionality which can either use the
performance clone or the model itself to do its work.