JackConnectionHandler: New class for handling jack (dis)connects in bundles
This will allow us to more gracefully perform the autoconnect work,
without doing a whole bunch of disconnections that then turn back into
connections. The logic here is that we simply collect all of the states
that we want to end up with, and then at the end commit the whole bunch
of all the changes, so we end up only actually making the changes that
are changes, and not disconnecting things we want to end up connected.