SamplerSynthVoice: When looping, don't wrap the loop data if we're crossfading
Since crossfading is an explicit choice, actually operate in a way that
makes that crossover point sound as expected (so, cross fade between the
two streams). This basically goes with the previous commit to ensure we
are phase correct, in making certain we sound as expected at the loop
points.