SequenceModel: Use hasContent instead of hasNotes for pattern saving
Previously, we would lose settings made on a pattern if you restarted
(for any reason), which, since we hold more than basic note lists is
arguably a destrucive action. So, we now use hasContent instead, and
will only not save the pattern if that is false, instead of only saving
when there are notes.