MidiRouter: Honour the key/scale locking options
This may seem somewhat roundabout (and it kind of is), but the intention
is to have as little of the logic as possible in the process loop. That
means we need to update a set of cached settings on the track from
somewhere else, and the best place to do that is in the cuia feedback
event function, and by calling that update function when the patterns
themselves change.