MidiRouterDevice: Add completeInitialisation(), also visible property
The visible property is just exposed so the devices model can give that
to the views and whatnot (or be used for filtering).

The new completeInitialisation() function replaces the hacky timer
completion with an explicit completion function. It is less proper (in
that it requires a user of the class to know to call that function,
which i guess we could guard for), but also, these are created and
destroyed by MidiRouter, and while technically you could create one
without, that was never intended, and so, this is cleaner, and makes it
much simpler to actually use it. Tradeoffs are tradeoffs, and this is
betterer.