Slices: Add an exclusivity concept (when starting playback, any ongoing will be stopped)
This introduces the playback exclusivity concept to our slices, where
you can set any slice on a track to have some exclusivity group, and
if any other slice playing on the same track shares that group, playback
of that will be stopped immediately. This might also be called a "choke
group" in other places, most notably older sampling drum machines.