JackPassthrough: Switch from linear gain to GainHandler, set up clients properly
This lets us have individual gain limits for the various passthrough
clients, and in particular
- a -24 to +24 dB limit for the synth clients, to match the samples
- a -40 to +20 limit for the track and master clients, to match current mixer layout

This also gives us quite a few lovely side effects. In particular, synth
slots can now be visually much more similar to the sample slots, which
becomes important as those are about to be on the same page. It's also
just nice to be able for both our sound sources to be the same anyway.