MidiRouter: Expose the synctimer devices' target tracks
This is a subset of some of the other things going on, but it allows us
to handle a few things more gracefully.

While working on this, i'm noticing we've got the track devices as well,
and... i'm wondering if we actually still need those? Thinking we likely
don't, but leaving them in just now, while i'm working on this.