Initial work on SysEx support
This is a pretty large pile of work, and it was starting to feel a
little silly to not commit it. So, while this is doubtless going to go
through a great many changes, here's the initial dump. The classes
themselves lack proper documentation, but each function in them is
reasonably documented.

The core access point is the new sysexHelper property on
MidiRouterDevice, which will give you access to an always-available
instance of SysexHelper (clever, i know), through which you can then
perform the various required actions, such as creating and sending SysEx
messages to the device.

Handling of incoming sysex is currently not amazingly well exposed, and
essentially just used for sniffing some of the known messages
(especially and in particular the Universal SysEx Identity Response type
of message, used to create the SysexIdentity instance found on
SysexHelper in the identity property).

SysexIdentity is fairly incomplete for now and will so far just split
out the bytes, and identify *some* manufacturers (not all, the list in
SysexIdTable.h is not yet complete), and none of the others.

SysexMessage is fairly complete as documented, but will likely need a
whole pile more helper functions to make life easier for any users of
the functionality in general, and also does not yet produce the checksum
when that setting is enabled.