SketchpadTrackInfo: Add an observable track class
This means we don't have to use those two (now deprecated) functions to
fetch the controller objects. We can also add further information on
this new class as we want to, but for now, this helps us correctly
update the external device on the track without having to do weird
hoop-jumping.