SndLibrary : Update implementation to base on symlinks instead of stats file
The idea is to index the snd files based on the categories using symlinks instead of using a stats file. All the snd files that are process are given a fileIdentifier in the format `<origin>/<username in case of community sounds>/<snd file name>` which is the path to the snd file relative to `/zynthian/zynthian-my-sounds/sounds/`. The file identifier is base64encoded and used as the symlink file name when indexed. The symlink is created here :  `${ZYNTHBOX_SND_INDEX_PATH}/<category id>/<fileIdentifier>`

The model lists the index directory and populates the model by reading the name from the symlink target and using the dir name as the category.