SamplerSynth: Hold list of CAS internally, use for off notes, and auto-delete unhandled clip commands
This is a first step towards ensuring that we operate more like a proper
synth engine on the tracks where we are supposed to do that. It struck
me while working on this that the intended functionality of the sample
picking style on tracks simply doesn't work correctly, which results in
it being simultaneously a bit weird, and also just kind of hard to
understand. It seems like it would be much more reasonable to simply
employ the key split logic for that, which would allow for similar
things, but also more flexibility in general. More thought required,
but... it seems like that would make sense (it would also make
SamplerSynth code simpler again, and just use the one location for the
clip command generation)