SamplerSynth: Pull all sample picking into SamplerSynth directly
This removes the need for asking the PlayGridManager to ask the
PatternModel instances to please hand back whatever samples have been
registered on a track for playback purposes and whatnot. Which was
always just kind of not really great anyway.

We now handle all this internally in each SamplerChannel, which also
keeps track of the order of samples, ensuring that we put them in the
correct order (for when in "first" mode), and also then use midi events'
midi channel to decide which samples to play when in "same" mode. All in
all, this is much cleaner than the old system, and consequently also
more easy to deal with for future additions and whatnot.

It might very well also be more performant, not at all sure about that.
It feels much nicer, though, whatever else is going on, so that's
definitely a bonus.