PatternModel: Clean up sample handling, default note length handling
This firstly and most importantly removes the no longer needed handling
of the various sample bits that used to live in PatternModel, called by
SamplerSynth via PlayGridManager. With it gone from here, there are a
number of other things which could likely also be removed, but for now
the core functions have, at least, been moved over to SamplerSynth
itself, and all PatternModel has to really deal with now is, in essence,
midi events. Much cleaner.

Default note length handling implements the concept in the ui which
allows us to set the note length to -1 to select the current default
length as set on the pattern (which by default is the step length,
resulting in identical behaviour to length 0, but which will change if
the default length is set to something else).