ProcessWrapper: Block the blocking call via a mutex
This is extremely serious business, and when we say "blocking", really
mean it.

You will still be allowed to send to the process, so if you are using
call, probably *only* use call and not the other options... If this
works, though, it would likely make sense to also honour the block when
performing general purpose sends.