ProcessWrapper: Add a way to detect calls in progress
This is a method using which we can refuse to operate multiple calls
simmultaneously. It is also not at all what i wanted to do, but i simply
cannot work out how to make KPty able to read from the tty without
having the event loop running as well. I don't see any good reason why
that would be the case, but it clearly *is* the case, so this is as
close as i can work out how to get it.

It puts a work burden on the consumer of the functionality, which is
definitely not what i wanted for something that is supposed to be
convenience functionality, but i also cannot work out what else to do
here.