ProcessWrapper: Rework ProcessWrapper to a serial transaction based approach
This approach should ensure that we can handle multiple calls to do
things which happen on top of each other, without causing the kind of
prroblem we used to have.

This currently does *not* support simply sending something and then
ignoring whether there's something going on (this would be fairly
trivial to add, and may well be needed for some of the engines).