ProcessWrapper: Reintroduce the expected output logic, plus be more blocky
We previously had the ability to wait for specific output, instead of
waiting for only the given command prompt. This introduces a new approach
to that same logic, which allows us to have both of the cases
(essentially, we will still want a command prompt, but any transaction can
be given a specific thing to look for instead of that prompt). One
corner case is arguably the "don't wait for anything" case, but with the
sequential system we have, probably just pass a \n for that, since that
would be automatically appended to anything sent to the system anyway.

If we need to be able to send raw data, we will need an explicit way to
do that, which will need to operate in a less hand-holdy fashion than
the transaction system (but without getting in the way of it as well).
For now, though, not on the list.

This isn't a guaranteed thing, but it feels like maybe this works
without doing the kind of weird behaviour it did previously where we'd
end up blocking forever on some things... Hopefully this works better.