Add a modified version of FolderListModel which supports # in folder names
It does this by testing for local files first before doing the big
decoding work, and then if it is a local file, just trust it's a path...
because that's what we do. This basically is not upstreamable, but yeah,
we need the ability, so this will have to do for us. Very annoying bug
for sure, though.