MidiRouterDeviceModel: Add "same" track, add sample slots as sources, change track to be synth
The "same" track is a logical "the track the destination is on", which
is useful for things like side-chain ducking between slots on the same
track (for a simple example: creating a techno rumbler)

This leaves us more space for future expansion (like being able to
actually use the post-fader output of a track as a source). We also now
have sample slots as a source option.