JackPassthrough: Add optional wetInput ports, and optional, default-off port creation
The first switches the logic around for dry/wet mixing, which previously
was a post-eq/comp effect. When a wet input port exists, we will instead
mix the dry and wet inputs down, and *then* apply eq/comp, before
writing that output to all enabled outputs. This allows the client to
function both as our track passthrough (where we need to output to the
two send effects), and as an insert effect handler (where we want to
apply the mix/eq/comp to the incoming signal rather than the outgoing).

The second part allows us to not create (and process) passthrough
clients until they're actually needed. This greatly reduces the process
cost of our mixerbus (for the default set on a standard rpi5, it went
from about 16% to 17% DSP saturation to around 6%).