JackConnectionHandler: Adjust logic a touch, so we only fetch ports when needed
That is, we don't fetch the ports until *right* at the end (which also
just happens to allow us to create and remove ports as we go along, and
ensures we aren't trying to get information on ports which have either
not been created yet, or which have been removed).

Also simply allow disconnections to fail, if one or both ports are
missing, because well... that kind of means the disconnection worked, as
that happens on removal of the port anyway.