JPVisualiserItem: Add support for AudioLevels, and a couple of visual control options
AudioLevels now also has the eq/comp loveliness, so that wants to be
supported here, so we can visualise it properly.

We also want to be able to have some much smaller variants of the
visualisation, which doesn't quite expose everything (only really what's
active and such), so adding in a couple of options to make it possible
to control that a bit.