Add external clock source support to the midi routing system and SyncTimer
Allow the user to set a specific external device as the clock source,
which in turn allows that external device to control the effective BPM
of the playback system. Only one device will be allowed at any one time,
and it is trusted explicitly to supply clock (which basically would
result in the system failing to do playback, and also set proper bpm for
any effect engines, if the clock signal is not supplied).

This should have been more smaller commits, but it is fairly intricate
and interconnected, so each of those commits would not have been
individually compileable, so... instead, single monolithic commit. The
system is as-yet untested, but it also should not affect anything until
you actually set a device as clock source.

We do not currently lock the ppqn for the devices, but this will need to
happen.