CASSliceSettings: Add a play/stop pair for testing purposes
This will initially be used for the step buttons when in slot mode, but
it seems like it'd probably be generally useful as well. Might need a
fire-and-forget type test thing as well for these, but... that'll need a
bit more thought to make work properly for all playback styles.