SamplerSynthVoice: Handle aftertouch a bit more gracefully
Basically, simply trusting aftertouch doesn't make sense (as that will
jump up all manner of ugly), so what you need to be able to do is
increase the value, but not *decrease* it, otherwise we end up jumping
too hard when performing using aftertouch. This feels like it might need
to be a little more elaborate, but this is a reasonable tradeoff at this
point, and we can look at the reduction later on (thought experiment on
that would be, allow user to first reach that point once, and *then*
allow reducing the amount after actually reaching the initial gain
value, which is sort of how many "catch the value before changing" type
interaction works with many midi things).