SyncTimer: Add scheduling function for controller buffers
While the Controller is usually used to immediately send stuff out, we
also want to use this buffer for test-firing sequencer steps (to avoid
potential clashes during playback, because pretty obviously people are
going to be doing that). It's also just more sematically correct that it
should be done by a controller rather than the sequencer.