PGM/SequenceModel: Perform a stop action on all notes off, which resets the patterns
This ensure that if you hit stop while not playing, the probability
settings of the patterns gets reset appropriately. This in turn ensures
that test-firing steps with probabilities outside of playback will
operate correctly (that is, the probability sequences will play
correctly), and that we can still reset things before playback, without
doing so implicitly.