PatternModel: Add step transposition function
This allows you to easily and safely transpose either all the entries on
a step, or a specific entry.

Future work would likely want to be allowing for similarly safe (because
that's fairly straightforward) transposition of a selection of entries,
instead of only a single one, but that's going to have to be a separate
function.